+++ apps/api/src/app/admin/admin.service.ts
@@ -23,6 +23,7 @@ import {
   AdminMarketData,
   AdminMarketDataDetails,
   AdminMarketDataItem,
+  AdminUserResponse,
   AdminUsersResponse,
   AssetProfileIdentifier,
   EnhancedSymbolProfile,
@@ -35,7 +36,8 @@ import {
   BadRequestException,
   HttpException,
   Injectable,
-  Logger
+  Logger,
+  NotFoundException
 } from '@nestjs/common';
 import {
   AssetClass,
@@ -507,6 +509,18 @@ export class AdminService {
     };
   }
 
+  public async getUser(id: string): Promise<AdminUserResponse> {
+    const [user] = await this.getUsersWithAnalytics({
+      where: { id }
+    });
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
+    return user;
+  }
+
   public async getUsers({
     skip,
     take = Number.MAX_SAFE_INTEGER
@@ -516,7 +530,15 @@ export class AdminService {
   }): Promise<AdminUsersResponse> {
     const [count, users] = await Promise.all([
       this.countUsersWithAnalytics(),
-      this.getUsersWithAnalytics({ skip, take })
+      this.getUsersWithAnalytics({
+        skip,
+        take,
+        where: {
+          NOT: {
+            analytics: null
+          }
+        }
+      })
     ]);
 
     return { count, users };
@@ -814,16 +836,16 @@ export class AdminService {
 
   private async getUsersWithAnalytics({
     skip,
-    take
+    take,
+    where
   }: {
     skip?: number;
     take?: number;
+    where?: Prisma.UserWhereInput;
   }): Promise<AdminUsersResponse['users']> {
     let orderBy: Prisma.Enumerable<Prisma.UserOrderByWithRelationInput> = [
       { createdAt: 'desc' }
     ];
-
-    let where: Prisma.UserWhereInput;
 
     if (this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION')) {
       orderBy = [
@@ -833,12 +855,6 @@ export class AdminService {
           }
         }
       ];
-
-      where = {
-        NOT: {
-          analytics: null
-        }
-      };
     }
 
     const usersWithAnalytics = await this.prismaService.user.findMany({